+++ .gitignore
@@ -25,7 +25,7 @@ yarn-debug.log*
 yarn-error.log*
 
 # local env files
-.env*.local
+.env*
 
 # vercel
 .vercel
@@ -35,5 +35,3 @@ yarn-error.log*
 next-env.d.ts
 
 venv
-
-.env.local
