+++ next.config.mjs
@@ -1,0 +1,10 @@
+// @ts-check
+
+/**
+ * @type {import('next').NextConfig}
+ */
+const nextConfig = {
+  output: 'standalone',
+};
+
+export default nextConfig;
