+++ package.json
@@ -7,13 +7,10 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "build-staging": "cross-env NODE_ENV=staging next build",
-    "start:staging": "cross-env NODE_ENV=staging next start",
     "lint": "next lint",
     "format": "prettier --write . --ignore-path .gitignore",
     "test:jest": "jest --ci",
     "test:jest-watch": "jest --watchAll",
-    "dev:server": "json-server --watch sessions_db.json --port 3001",
     "test:cypress": "cypress open"
   },
   "eslintConfig": {
@@ -38,7 +35,7 @@
     "printWidth": 100
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.4.2",
+    "@hookform/resolvers": "^3.6.0",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-checkbox": "^1.0.4",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
@@ -48,13 +45,13 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.3",
     "@tanstack/react-table": "^8.17.3",
-    "aws-sdk": "^2.1624.0",
-    "axios": "^1.7.1",
+    "aws-sdk": "^2.1638.0",
+    "axios": "^1.7.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
-    "lucide-react": "^0.379.0",
-    "next": "14.2.3",
+    "lucide-react": "^0.394.0",
+    "next": "^14.2.3",
     "react": "18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "18.3.1",
@@ -64,25 +61,24 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.4.5",
-    "@testing-library/react": "^15.0.7",
+    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.12",
-    "@types/node": "^20.12.12",
-    "@types/react": "^18.3.2",
+    "@types/node": "^20.14.2",
+    "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "autoprefixer": "^10.4.19",
     "cross-env": "^7.0.3",
-    "cypress": "^13.9.0",
+    "cypress": "^13.11.0",
     "eslint": "^8.56.0",
     "eslint-config-next": "14.2.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "json-server": "^1.0.0-beta.0",
     "postcss": "^8.4.38",
     "pre-commit": "^1.2.2",
-    "prettier": "^3.2.5",
-    "tailwindcss": "^3.4.3",
+    "prettier": "^3.3.2",
+    "tailwindcss": "^3.4.4",
     "ts-node": "^10.9.2",
     "typescript": "^5.4.5"
   }
