+++ src/services/services.ts
@@ -1,4 +1,7 @@
+'use server';
+
 import { DATA_PER_PAGE } from '@/Constants';
+import { publishMessage } from '@/services/Aws';
 import { Session } from '@/types/api.types';
 import { instance } from '../lib/axios';
 
@@ -25,4 +28,43 @@ export async function getASession(id: number | null) {
 
   const { data } = await instance.get<Session>(`/session/${id}`);
   return data;
+}
+
+export async function createSession(formData: FormData) {
+  try {
+    const response = await instance.post(`/api/session`, {});
+
+    const sessionId = response.data.id;
+    const message = {
+      action: 'db_id',
+      id: sessionId,
+    };
+
+    publishMessage(message);
+
+    return {
+      id: sessionId,
+    };
+  } catch (error) {
+    console.error('Error posting form data', error);
+    throw error;
+  }
+}
+
+export async function updateSession(formData: FormData) {
+  try {
+    const message = {
+      action: 'patch',
+      // id: sessionId,
+      // patch_session: patchBody,
+    };
+    publishMessage(message);
+
+    return {
+      // id: sessionId,
+    };
+  } catch (error) {
+    console.error('Error posting form data', error);
+    throw error;
+  }
 }
